requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
PyMySQL>=1.1.0
cryptography>=41.0.7
sqlalchemy>=2.0.23
alembic>=1.12.1
qdrant-client>=1.7.1
python-multipart>=0.0.6
ollama>=0.1.7
aiohttp>=3.9.0
aiofiles>=23.2.0
pdfminer3k>=1.3.4
tiktoken>=0.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
